Smooth ISO: fine-tuning for ISO 6400 (60D and others)

+++ platform/60D.111/consts.h
@@ -319,7 +319,7 @@
 #define DISPLAY_IS_ON (DISPLAY_STATEOBJ->current_state != 0)
 
 #define VIDEO_PARAMETERS_SRC_3 0x4FDA8
-#define FRAME_ISO (*(uint16_t*)(VIDEO_PARAMETERS_SRC_3+0x8))
+#define FRAME_ISO (*(uint8_t*)(VIDEO_PARAMETERS_SRC_3+0x8))
 #define FRAME_SHUTTER (*(uint8_t*)(VIDEO_PARAMETERS_SRC_3+0xa))
 #define FRAME_BV (*(uint8_t*)(VIDEO_PARAMETERS_SRC_3+0xb))
